refactor(client directory structure): refactored client's directory structure
