+++ client/src/components/Clock/hooks/useClock.hook.ts
