+++ client/src/components/SensorList/queries/useSensorListUpdatesSubscription.query.ts
