+++ client/src/components/Weather/components/MultiDay/MultiDay.component.tsx
