+++ client/src/components/Weather/queries/useAirQuality.query.ts
