+++ client/src/components/Weather/queries/useWeather.query.ts
