+++ client/src/components/Weather/utils/loadWeatherIcons.util.ts
