+++ client/src/components/Weather/utils/mapThresholdToColor.util.ts
