+++ client/src/components/Weather/utils/mapWeatherCodeToImage.util.ts
